changed fonts and deleted unused code in User.java
