+++ .idea/misc.xml
@@ -3,6 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="app/src/main/res/drawable/btn_custom.xml" value="0.1155" />
+        <entry key="app/src/main/res/drawable/et_custom.xml" value="0.1155" />
         <entry key="app/src/main/res/layout/action_view_progress.xml" value="0.40260416666666665" />
         <entry key="app/src/main/res/layout/activity_login.xml" value="0.2203125" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.375" />
@@ -11,7 +13,7 @@
         <entry key="app/src/main/res/layout/activity_stream_details.xml" value="0.2703125" />
         <entry key="app/src/main/res/layout/activity_stream_home_details.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/activity_stream_homes.xml" value="0.2171875" />
-        <entry key="app/src/main/res/layout/dropdown_item.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/dropdown_item.xml" value="0.335" />
         <entry key="app/src/main/res/layout/fragment_login.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_preference.xml" value="0.39386138916015623" />
         <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.3964102564102564" />
