+++ app/src/main/res/font/pt_sans.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<font-family xmlns:app="http://schemas.android.com/apk/res-auto"
+        app:fontProviderAuthority="com.google.android.gms.fonts"
+        app:fontProviderPackage="com.google.android.gms"
+        app:fontProviderQuery="PT Sans"
+        app:fontProviderCerts="@array/com_google_android_gms_fonts_certs">
+</font-family>
