+++ app/src/main/res/layout/activity_login.xml
@@ -15,8 +15,8 @@
         android:layout_marginStart="123dp"
         android:layout_marginTop="30dp"
         android:layout_marginEnd="147dp"
-        android:fontFamily="@font/quando"
-        android:text="@string/app_name"
+        android:fontFamily="@font/pt_sans_bold"
+        android:text="@string/login"
         android:textSize="50sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -29,7 +29,7 @@
         android:layout_above="@id/EmailEditText"
         android:layout_marginStart="136dp"
         android:layout_marginEnd="165dp"
-        android:fontFamily="@font/calistoga"
+        android:fontFamily="@font/pt_sans_bold"
         android:text="@string/tag"
         android:textAlignment="center"
         android:textSize="20sp"
@@ -101,6 +101,7 @@
         android:layout_centerHorizontal="true"
         android:layout_marginStart="60dp"
         android:layout_marginTop="16dp"
+        android:fontFamily="@font/pt_sans"
         android:text="@string/swipe_left_for_register"
         android:textSize="15sp"
         app:layout_constraintEnd_toStartOf="@+id/ArrowImageView"
