+++ app/src/main/res/layout/activity_register.xml
@@ -11,12 +11,13 @@
         android:id="@+id/LogoTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="25dp"
+        android:layout_marginTop="60dp"
+        android:layout_marginBottom="5dp"
+        android:fontFamily="@font/pt_sans_bold"
         android:text="@string/register"
         android:textAlignment="center"
-        android:textSize="50sp"
-        android:layout_marginStart="25dp"
-        android:layout_marginBottom="5dp"
-        android:layout_marginTop="60dp"/>
+        android:textSize="50sp" />
 
     <TextView
         android:id="@+id/SubtitleTextView"
@@ -24,7 +25,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="25dp"
         android:layout_marginBottom="50dp"
-        android:fontFamily="@font/aguafina_script"
+        android:fontFamily="@font/pt_sans"
         android:text="@string/tag"
         android:textSize="20sp" />
 
@@ -100,6 +101,7 @@
             android:layout_alignParentBottom="true"
             android:layout_centerHorizontal="true"
             android:layout_marginBottom="20dp"
+            android:fontFamily="@font/pt_sans"
             android:text="@string/swipe_right_for_login"
             android:textSize="15sp" />
