+++ app/src/main/res/layout/dropdown_item.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-    <TextView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/textView"
-        android:layout_width="match_parent"
-        android:textColor="@color/black"
-        android:textStyle="bold"
-        android:padding="14dp"
-        android:layout_height="wrap_content"
-        android:text="TextView" />
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/textView"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:fontFamily="@font/pt_sans_bold"
+    android:padding="14dp"
+    android:text="TextView"
+    android:textColor="@color/black"
+    android:textStyle="bold" />
