+++ app/src/main/res/layout/fragment_profile.xml
@@ -21,8 +21,9 @@
         android:id="@+id/ProfileUsernameTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
+        android:layout_marginStart="48dp"
         android:layout_marginTop="16dp"
+        android:fontFamily="@font/pt_sans_bold"
         android:text="TextView"
         android:textSize="20sp"
         app:layout_constraintStart_toEndOf="@id/ProfilePicImageView"
@@ -44,10 +45,23 @@
         android:id="@+id/ProfileEmailTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
+        android:layout_marginStart="48dp"
         android:layout_marginTop="3dp"
+        android:fontFamily="@font/pt_sans_bold"
         android:text="TextView"
         android:textSize="20sp"
         app:layout_constraintStart_toEndOf="@+id/ProfilePicImageView"
         app:layout_constraintTop_toBottomOf="@+id/ProfileUsernameTextView" />
+
+    <Button
+        android:id="@+id/LogoutButton"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
+        android:layout_marginTop="20dp"
+        android:text="LOGOUT"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/ProfileAddProfilePicImageView"
+        tools:layout_conversion_absoluteHeight="48dp"
+        tools:layout_conversion_absoluteWidth="93dp" />
 </androidx.constraintlayout.widget.ConstraintLayout>
