+++ app/src/main/res/layout/item_home.xml
@@ -20,6 +20,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
+        android:fontFamily="@font/pt_sans"
         android:text="TextView"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
         app:layout_constraintTop_toBottomOf="@+id/BedroomText" />
@@ -29,6 +30,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
+        android:fontFamily="@font/pt_sans"
         android:text="TextView"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
         app:layout_constraintTop_toBottomOf="@+id/YearBuiltText" />
@@ -38,6 +40,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
+        android:fontFamily="@font/pt_sans"
         android:text="TextView"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
         app:layout_constraintTop_toBottomOf="@+id/PropertyTypeText" />
@@ -48,6 +51,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
         android:layout_marginTop="4dp"
+        android:fontFamily="@font/pt_sans_bold"
         android:text="TextView"
         android:textSize="10.5sp"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
@@ -59,7 +63,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
         android:layout_marginTop="1dp"
-        android:fontFamily="More Fonts..."
+        android:fontFamily="@font/pt_sans"
         android:text="TextView"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
         app:layout_constraintTop_toBottomOf="@+id/AddressTextView" />
@@ -70,6 +74,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
         android:layout_marginTop="1dp"
+        android:fontFamily="@font/pt_sans"
         android:text="TextView"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
         app:layout_constraintTop_toBottomOf="@+id/DistanceTextView" />
@@ -79,6 +84,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="1dp"
+        android:fontFamily="@font/pt_sans"
         android:text="TextView"
         app:layout_constraintStart_toEndOf="@+id/HomeImageView"
         app:layout_constraintTop_toBottomOf="@+id/BathroomText" />
