+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -2,12 +2,6 @@
 <menu xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/action_profile"
-        android:enabled="true"
-        android:icon="@android:drawable/ic_menu_view"
-        android:title="Profile"
-        app:showAsAction="always" />
-    <item
         android:id="@+id/action_preferences"
         android:enabled="true"
         android:icon="@android:drawable/ic_menu_edit"
@@ -18,5 +12,11 @@
         android:enabled="true"
         android:icon="@android:drawable/ic_menu_mapmode"
         android:title="Stream"
+        app:showAsAction="always" />
+    <item
+        android:id="@+id/action_profile"
+        android:enabled="true"
+        android:icon="@android:drawable/ic_menu_view"
+        android:title="Profile"
         app:showAsAction="always" />
 </menu>
