+++ app/src/main/res/values/preloaded_fonts.xml
@@ -3,6 +3,8 @@
     <array name="preloaded_fonts" translatable="false">
         <item>@font/aguafina_script</item>
         <item>@font/calistoga</item>
+        <item>@font/pt_sans</item>
+        <item>@font/pt_sans_bold</item>
         <item>@font/quando</item>
         <item>@font/rammetto_one</item>
     </array>
